完成菜式信息接口,菜式成分接口

+++ src/main/java/com/bkit/fatdown/service/impl/FoodElementServiceImpl.java
@@ -128,8 +128,7 @@ public class FoodElementServiceImpl implements IFoodElementService {
         TbFoodElementRelationExample example = new TbFoodElementRelationExample();
         example.createCriteria()
                 .andIdEqualTo(relationId);
-
-        return relationMapper.countByExample(example);
+        return (int)relationMapper.countByExample(example);
     }
 
     /**
